refactor: Move addAccountPair and restore functions into Keyring class file
