--- packages/ui-keyring/src/account/addPair.ts
@@ -1,15 +1,0 @@
-// Copyright 2017-2018 @polkadot/ui-keyring authors & contributors
-// This software may be modified and distributed under the terms
-// of the ISC license. See the LICENSE file for details.
-
-import { KeyringPair$Json } from '@polkadot/util-keyring/types';
-import { State } from '../types';
-
-export default function addAccountPair (state: State, json: KeyringPair$Json): void {
-  if (!json.meta.whenCreated) {
-    json.meta.whenCreated = Date.now();
-  }
-
-  state.keyring.addFromJson(json);
-  state.accounts.add(json.address, json);
-}
