+++ packages/ui-keyring/src/index.ts
@@ -6,17 +6,18 @@ import { KeyringPair, KeyringPair$Meta, KeyringPair$Json } from '@polkadot/util-
 import { SingleAddress } from './observable/types';
 import { KeyringAddress, KeyringInstance, State } from './types';
 
+import hexToU8a from '@polkadot/util/hex/toU8a';
+import createPair from '@polkadot/util-keyring/pair';
+import decodeAddress from '@polkadot/util-keyring/address/decode';
 import testKeyring from '@polkadot/util-keyring/testing';
 
 import accounts from './observable/accounts';
 import addresses from './observable/addresses';
 import development from './observable/development';
 import loadAll from './loadAll';
-import addPair from './account/addPair';
 import backupAccount from './account/backup';
 import createAccount from './account/create';
 import forgetAccount from './account/forget';
-import restoreAccount from './account/restore';
 import isAvailable from './isAvailable';
 import isPassValid from './isPassValid';
 import saveAccount from './account/save';
@@ -41,8 +42,13 @@ class Keyring implements KeyringInstance {
     this.loadAll();
   }
 
-  addPair (json: KeyringPair$Json): void {
-    return addPair(this.state, json);
+  addAccountPair (json: KeyringPair$Json): void {
+    if (!json.meta.whenCreated) {
+      json.meta.whenCreated = Date.now();
+    }
+
+    this.state.keyring.addFromJson(json);
+    this.state.accounts.add(json.address, json);
   }
 
   backupAccount (pair: KeyringPair, password: string): KeyringPair$Json {
@@ -96,7 +102,21 @@ class Keyring implements KeyringInstance {
   }
 
   restoreAccount (json: KeyringPair$Json, password: string): KeyringPair {
-    return restoreAccount(this.state, json, password);
+    const pair = createPair(
+      {
+        publicKey: decodeAddress(json.address),
+        secretKey: new Uint8Array()
+      },
+      json.meta,
+      hexToU8a(json.encoded)
+    );
+
+    pair.decodePkcs8(password);
+    pair.lock();
+    this.state.keyring.addPair(pair);
+    this.addAccountPair(json);
+
+    return pair;
   }
 
   saveAccount (pair: KeyringPair, password?: string): void {
